Revert "Merge pull request #8906 from amcasey/HidePane2"
This reverts commit 0e7a2b6e4fd3e79a78843319af965f5153bf21ad, reversing
changes made to e1f34faa3a7f8c60c9489cfee0d46d918cfe0652.
