--- src/InteractiveWindow/VisualStudio/PublicAPI.Unshipped.txt
@@ -1,4 +1,0 @@
-Microsoft.VisualStudio.InteractiveWindow.Shell.IVsInteractiveWindow2
-Microsoft.VisualStudio.InteractiveWindow.Shell.IVsInteractiveWindow2.WindowFrame.get -> object
-Microsoft.VisualStudio.InteractiveWindow.Shell.IVsInteractiveWindow2.BitmapImageMoniker.get -> Microsoft.VisualStudio.Imaging.Interop.ImageMoniker
-Microsoft.VisualStudio.InteractiveWindow.Shell.IVsInteractiveWindow2.BitmapImageMoniker.set -> void
