+++ src/InteractiveWindow/VisualStudio/VisualStudioInteractiveWindow.csproj
@@ -53,9 +53,6 @@
     <Reference Include="Microsoft.VisualStudio.Editor, Version=$(VisualStudioReferenceAssemblyVersion), Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a">
       <Private>false</Private>
     </Reference>
-    <Reference Include="Microsoft.VisualStudio.Imaging.Interop.14.0.DesignTime, Version=14.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a">
-      <EmbedInteropTypes>True</EmbedInteropTypes>
-    </Reference>
     <Reference Include="Microsoft.VisualStudio.Language.Intellisense, Version=$(VisualStudioReferenceAssemblyVersion), Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a">
       <Private>false</Private>
     </Reference>
@@ -128,8 +125,6 @@
     <Compile Include="AssemblyRedirects.cs" />
     <Compile Include="CommandIds.cs" />
     <Compile Include="Guids.cs" />
-    <Compile Include="IVsInteractiveWindow2.cs" />
-    <Compile Include="VsInteractiveWindowPane.cs" />
     <Compile Include="VsInteractiveWindowEditorFactoryService.cs" />
     <Compile Include="InteractiveWindowPackage.cs" />
     <Compile Include="IVsInteractiveWindow.cs" />
